chore: tidy up test folder structure;
