+++ test/cases/sdk_test.go
@@ -1,173 +1,29 @@
-package otelchi
+package otelchi_test
 
 import (
-	"bufio"
-	"context"
-	"io"
-	"net"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
 	"github.com/go-chi/chi/v5"
+	"github.com/riandyrn/otelchi"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/propagation"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/sdk/trace/tracetest"
 	"go.opentelemetry.io/otel/trace"
 )
 
-var sc = trace.NewSpanContext(trace.SpanContextConfig{
-	TraceID:    [16]byte{1},
-	SpanID:     [8]byte{1},
-	Remote:     true,
-	TraceFlags: trace.FlagsSampled,
-})
-
-func TestPassthroughSpanFromGlobalTracer(t *testing.T) {
-	var called bool
-	router := chi.NewRouter()
-	router.Use(Middleware("foobar"))
-	// The default global TracerProvider provides "pass through" spans for any
-	// span context in the incoming request context.
-	router.HandleFunc("/user/{id}", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		called = true
-		got := trace.SpanFromContext(r.Context()).SpanContext()
-		assert.Equal(t, sc, got)
-		w.WriteHeader(http.StatusOK)
-	}))
-
-	r := httptest.NewRequest("GET", "/user/123", nil)
-	r = r.WithContext(trace.ContextWithRemoteSpanContext(context.Background(), sc))
-	w := httptest.NewRecorder()
-
-	router.ServeHTTP(w, r)
-	assert.True(t, called, "failed to run test")
-}
-
-func TestPropagationWithGlobalPropagators(t *testing.T) {
-	defer func(p propagation.TextMapPropagator) {
-		otel.SetTextMapPropagator(p)
-	}(otel.GetTextMapPropagator())
-
-	prop := propagation.TraceContext{}
-	otel.SetTextMapPropagator(prop)
-
-	r := httptest.NewRequest("GET", "/user/123", nil)
-	w := httptest.NewRecorder()
-
-	ctx := trace.ContextWithRemoteSpanContext(context.Background(), sc)
-	otel.GetTextMapPropagator().Inject(ctx, propagation.HeaderCarrier(r.Header))
-
-	var called bool
-	router := chi.NewRouter()
-	router.Use(Middleware("foobar"))
-	router.HandleFunc("/user/{id}", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		called = true
-		span := trace.SpanFromContext(r.Context())
-		assert.Equal(t, sc, span.SpanContext())
-		w.WriteHeader(http.StatusOK)
-	}))
-
-	router.ServeHTTP(w, r)
-	assert.True(t, called, "failed to run test")
-}
-
-func TestPropagationWithCustomPropagators(t *testing.T) {
-	prop := propagation.TraceContext{}
-
-	r := httptest.NewRequest("GET", "/user/123", nil)
-	w := httptest.NewRecorder()
-
-	ctx := trace.ContextWithRemoteSpanContext(context.Background(), sc)
-	prop.Inject(ctx, propagation.HeaderCarrier(r.Header))
-
-	var called bool
-	router := chi.NewRouter()
-	router.Use(Middleware("foobar", WithPropagators(prop)))
-	router.HandleFunc("/user/{id}", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		called = true
-		span := trace.SpanFromContext(r.Context())
-		assert.Equal(t, sc, span.SpanContext())
-		w.WriteHeader(http.StatusOK)
-	}))
-
-	router.ServeHTTP(w, r)
-	assert.True(t, called, "failed to run test")
-}
-
-type testResponseWriter struct {
-	writer http.ResponseWriter
-}
-
-func (rw *testResponseWriter) Header() http.Header {
-	return rw.writer.Header()
-}
-func (rw *testResponseWriter) Write(b []byte) (int, error) {
-	return rw.writer.Write(b)
-}
-func (rw *testResponseWriter) WriteHeader(statusCode int) {
-	rw.writer.WriteHeader(statusCode)
-}
-
-// implement Hijacker
-func (rw *testResponseWriter) Hijack() (net.Conn, *bufio.ReadWriter, error) {
-	return nil, nil, nil
-}
-
-// implement Pusher
-func (rw *testResponseWriter) Push(target string, opts *http.PushOptions) error {
-	return nil
-}
-
-// implement Flusher
-func (rw *testResponseWriter) Flush() {
-}
-
-// implement io.ReaderFrom
-func (rw *testResponseWriter) ReadFrom(r io.Reader) (n int64, err error) {
-	return 0, nil
-}
-
-func TestResponseWriterInterfaces(t *testing.T) {
-	// make sure the recordingResponseWriter preserves interfaces implemented by the wrapped writer
-	router := chi.NewRouter()
-	router.Use(Middleware("foobar"))
-	router.HandleFunc("/user/{id}", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		assert.Implements(t, (*http.Hijacker)(nil), w)
-		assert.Implements(t, (*http.Pusher)(nil), w)
-		assert.Implements(t, (*http.Flusher)(nil), w)
-		assert.Implements(t, (*io.ReaderFrom)(nil), w)
-		w.WriteHeader(http.StatusOK)
-	}))
-
-	r := httptest.NewRequest("GET", "/user/123", nil)
-	w := &testResponseWriter{
-		writer: httptest.NewRecorder(),
-	}
-
-	router.ServeHTTP(w, r)
-}
-
-func ok(w http.ResponseWriter, _ *http.Request) {
-	w.WriteHeader(http.StatusOK)
-}
-
 func TestSDKIntegration(t *testing.T) {
-	sr := tracetest.NewSpanRecorder()
-	provider := sdktrace.NewTracerProvider()
-	provider.RegisterSpanProcessor(sr)
+	router, sr := newSDKTestRouter("foobar")
 
-	router := chi.NewRouter()
-	router.Use(Middleware("foobar", WithTracerProvider(provider)))
 	router.HandleFunc("/user/{id:[0-9]+}", ok)
 	router.HandleFunc("/book/{title}", ok)
 
 	r0 := httptest.NewRequest("GET", "/user/123", nil)
 	r1 := httptest.NewRequest("GET", "/book/foo", nil)
+
 	w := httptest.NewRecorder()
 	router.ServeHTTP(w, r0)
 	router.ServeHTTP(w, r1)
@@ -199,7 +55,7 @@ func TestSDKIntegrationWithFilters(t *testing.T) {
 	provider.RegisterSpanProcessor(sr)
 
 	router := chi.NewRouter()
-	router.Use(Middleware("foobar", WithTracerProvider(provider), WithFilter(func(r *http.Request) bool {
+	router.Use(otelchi.Middleware("foobar", otelchi.WithTracerProvider(provider), otelchi.WithFilter(func(r *http.Request) bool {
 		if r.URL.Path == "/live" || r.URL.Path == "/ready" {
 			return false
 		}
@@ -248,10 +104,10 @@ func TestSDKIntegrationWithChiRoutes(t *testing.T) {
 
 	router := chi.NewRouter()
 	router.Use(
-		Middleware(
+		otelchi.Middleware(
 			"foobar",
-			WithTracerProvider(provider),
-			WithChiRoutes(router),
+			otelchi.WithTracerProvider(provider),
+			otelchi.WithChiRoutes(router),
 		),
 	)
 	router.HandleFunc("/user/{id:[0-9]+}", ok)
@@ -291,10 +147,10 @@ func TestSDKIntegrationOverrideSpanName(t *testing.T) {
 
 	router := chi.NewRouter()
 	router.Use(
-		Middleware(
+		otelchi.Middleware(
 			"foobar",
-			WithTracerProvider(provider),
-			WithChiRoutes(router),
+			otelchi.WithTracerProvider(provider),
+			otelchi.WithChiRoutes(router),
 		),
 	)
 	router.HandleFunc("/user/{id:[0-9]+}", func(w http.ResponseWriter, r *http.Request) {
@@ -338,10 +194,10 @@ func TestSDKIntegrationWithRequestMethodInSpanName(t *testing.T) {
 
 	router := chi.NewRouter()
 	router.Use(
-		Middleware(
+		otelchi.Middleware(
 			"foobar",
-			WithTracerProvider(provider),
-			WithRequestMethodInSpanName(true),
+			otelchi.WithTracerProvider(provider),
+			otelchi.WithRequestMethodInSpanName(true),
 		),
 	)
 	router.HandleFunc("/user/{id:[0-9]+}", ok)
@@ -376,7 +232,7 @@ func TestSDKIntegrationWithRequestMethodInSpanName(t *testing.T) {
 
 func assertSpan(t *testing.T, span sdktrace.ReadOnlySpan, name string, kind trace.SpanKind, attrs ...attribute.KeyValue) {
 	assert.Equal(t, name, span.Name())
-	assert.Equal(t, trace.SpanKindServer, span.SpanKind())
+	assert.Equal(t, kind, span.SpanKind())
 
 	got := make(map[attribute.Key]attribute.Value, len(span.Attributes()))
 	for _, a := range span.Attributes() {
@@ -388,4 +244,21 @@ func assertSpan(t *testing.T, span sdktrace.ReadOnlySpan, name string, kind trac
 		}
 		assert.Equal(t, got[want.Key], want.Value)
 	}
+}
+
+func ok(w http.ResponseWriter, _ *http.Request) {
+	w.WriteHeader(http.StatusOK)
+}
+
+func newSDKTestRouter(serverName string, opts ...otelchi.Option) (chi.Router, *tracetest.SpanRecorder) {
+	spanRecorder := tracetest.NewSpanRecorder()
+	tracerProvider := sdktrace.NewTracerProvider()
+	tracerProvider.RegisterSpanProcessor(spanRecorder)
+
+	opts = append(opts, otelchi.WithTracerProvider(tracerProvider))
+
+	router := chi.NewRouter()
+	router.Use(otelchi.Middleware(serverName, opts...))
+
+	return router, spanRecorder
 }
