+++ test/infras/install_go.sh
